+++ Fun.Blazor/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Code cleanup and refactor
 - Support to get value for other type in AdaptiveForm
 - Do not increase index for emptyAttr or emptyNode
+- Support CE empty body for fsharp 9
 
 ## [4.0.12] - 2024-06-05